Website Builder/requirements.txt
python-dotenv>=1.0.0
apify-client>=1.6.0
gspread>=6.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
httpx>=0.27.0
html2text>=2024.2.26
pandas>=2.1.0
python-whois>=0.9.4
